editing the delete to not delete if any connected with session

+++ backend/src/main/java/com/spring/techpractica/Application/Admin/Technology/DeleteTechnology/DeleteTechnologyUseCase.java
@@ -1,5 +1,7 @@
 package com.spring.techpractica.Application.Admin.Technology.DeleteTechnology;
 
+import com.spring.techpractica.Core.Session.SessionRepository;
+import com.spring.techpractica.Core.Shared.Exception.ResourcesDuplicateException;
 import com.spring.techpractica.Core.System.Entity.System;
 import com.spring.techpractica.Core.System.SystemRepository;
 import com.spring.techpractica.Core.Technology.Entity.Technology;
@@ -13,12 +15,15 @@ import org.springframework.transaction.annotation.Transactional;
 public class DeleteTechnologyUseCase {
     private final TechnologyRepository technologyRepository;
 
-
+    private final SessionRepository sessionRepository;
     @Transactional
     public void execute(DeleteTechnologyCommand command) {
 
         Technology technology = technologyRepository.getOrThrowByID(command.technologyId());
 
+        if (!sessionRepository.getSessionsByTechnologyId(command.technologyId()).isEmpty()) {
+            throw new ResourcesDuplicateException("session with technology id " + command.technologyId() + " exists");
+        }
          technologyRepository.deleteById(technology.getId());
     }
 }
